+++ tests/phpunit/Scaler/FileSizeScalerTest.php
@@ -8,6 +8,7 @@ use Brendt\Image\Scaler\FileSizeScaler;
 use Intervention\Image\ImageManager;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Filesystem\Filesystem;
+use Symfony\Component\Finder\Finder;
 
 class FileSizeScalerTest extends TestCase
 {
@@ -54,29 +55,27 @@ class FileSizeScalerTest extends TestCase
     }
 
     public function test_scale_down() {
-        $responsiveImage = $this->createResponsiveImage();
+        $sourceFile = $this->createSourceFile();
         $imageObject = $this->createImageObject();
 
-        $responsiveImage = $this->scaler->scale($responsiveImage, $imageObject);
+        $sizes = $this->scaler->scale($sourceFile, $imageObject);
 
-        $this->assertTrue(count($responsiveImage->getSrcset()) > 1);
+        $this->assertTrue(count($sizes) > 1);
     }
 
     // TODO: test algorithm
-
-    private function createResponsiveImage() {
-        $responsiveImage = new ResponsiveImage('img/image.jpeg');
-        $responsiveImage->setExtension('jpeg');
-        $responsiveImage->setFileName('image');
-        $responsiveImage->setUrlPath('/img');
-
-        return $responsiveImage;
-    }
 
     private function createImageObject() {
         $imageObject = $this->engine->make('./tests/img/image.jpeg');
 
         return $imageObject;
+    }
+
+    private function createSourceFile() {
+        $sourceFiles = Finder::create()->files()->in('./tests/img')->name('image.jpeg')->getIterator();
+        $sourceFiles->rewind();
+
+        return $sourceFiles->current();
     }
 
 }